requirements.txt
boto3
botocore
paramiko
python-simplestreams
pyyaml
